+++ examples/vue/ssr/src/main.js
@@ -75,22 +75,13 @@ export async function createApp({
               if (typeof window !== 'object') {
                 return;
               }
-              this._onPopState = (event) => {
+              this._onPopState = () => {
                 if (this.writeTimer) {
                   window.clearTimeout(this.writeTimer);
                   this.writeTimer = undefined;
                 }
 
-                const routeState = event.state;
-
-                // At initial load, the state is read from the URL without update.
-                // Therefore the state object is not available.
-                // In this case, we fallback and read the URL.
-                if (!routeState) {
-                  callback(this.read());
-                } else {
-                  callback(routeState);
-                }
+                callback(this.read());
               };
 
               window.addEventListener('popstate', this._onPopState);