+++ packages/instantsearch.js/src/connectors/infinite-hits/connectInfiniteHits.ts
@@ -350,19 +350,25 @@ export default (function connectInfiniteHits<
 
         const banner = results?.renderingContent?.widgets?.banners?.[0];
 
-        if (!results) {
+        if (!showPrevious) {
           showPrevious = getShowPrevious(helper);
           showMore = getShowMore(helper);
+        }
+
+        if (!sendEvent) {
           sendEvent = createSendEventForHits({
             instantSearchInstance,
-            getIndex: () => helper.getIndex(),
+            helper,
             widgetType: this.$$type,
           });
           bindEvent = createBindEventForHits({
-            getIndex: () => helper.getIndex(),
+            helper,
             widgetType: this.$$type,
             instantSearchInstance,
           });
+        }
+
+        if (!results) {
           isFirstPage =
             state.page === undefined ||
             getFirstReceivedPage(state, cachedHits) === 0;