+++ packages/instantsearch.js/src/connectors/looking-similar/connectLookingSimilar.ts
@@ -166,7 +166,7 @@ export default (function connectLookingSimilar<
         if (!sendEvent) {
           sendEvent = createSendEventForHits({
             instantSearchInstance,
-            getIndex: () => helper.getIndex(),
+            helper,
             widgetType: this.$$type,
           });
         }