+++ packages/instantsearch.js/src/connectors/trending-items/connectTrendingItems.ts
@@ -187,7 +187,7 @@ export default (function connectTrendingItems<
         if (!sendEvent) {
           sendEvent = createSendEventForHits({
             instantSearchInstance,
-            getIndex: () => helper.getIndex(),
+            helper,
             widgetType: this.$$type,
           });
         }