+++ packages/react-instantsearch-core/src/lib/useSearchResults.ts
@@ -51,6 +51,11 @@ export function useSearchResults(): SearchResultsApi {
 
     search.addListener('render', handleRender);
 
+    // Force setting results to mitigate potential race conditions where
+    // render listener is added after search results have been returned.
+    // This edge case is currently not covered by the tests.
+    handleRender();
+
     return () => {
       search.removeListener('render', handleRender);
     };