+++ packages/react-instantsearch-nextjs/src/__tests__/InitializePromise.test.tsx
@@ -6,10 +6,10 @@ import {
   createSearchClient,
   createSingleSearchResponse,
 } from '@instantsearch/mocks';
-import { act, render } from '@testing-library/react';
 import * as utils from 'instantsearch.js/es/lib/utils';
 import { ServerInsertedHTMLContext } from 'next/navigation';
 import React from 'react';
+import { renderToString } from 'react-dom/server';
 import { SearchBox, TrendingItems } from 'react-instantsearch';
 import {
   InstantSearch,
@@ -27,7 +27,7 @@ jest.mock('instantsearch.js/es/lib/utils', () => ({
   resetWidgetId: jest.fn(),
 }));
 
-const renderComponent = ({
+const renderComponent = async ({
   children,
   ref = { current: null },
   nonce,
@@ -44,7 +44,7 @@ const renderComponent = ({
     }),
   });
 
-  render(
+  renderToString(
     <InstantSearchRSCContext.Provider value={ref}>
       <InstantSearchSSRProvider>
         <InstantSearch searchClient={client} indexName="indexName">
@@ -60,6 +60,8 @@ const renderComponent = ({
     </InstantSearchRSCContext.Provider>
   );
 
+  await ref.current;
+
   return client;
 };
 
@@ -70,17 +72,11 @@ test('it calls resetWidgetId', () => {
 });
 
 test('it applies provided nonce on the injected script tag', async () => {
-  const ref: { current: PromiseWithState<void> | null } = { current: null };
   const insertedHTML = jest.fn();
-  renderComponent({
-    ref,
+  await renderComponent({
     children: <SearchBox />,
     nonce: 'csp-nonce',
     insertedHTML,
-  });
-
-  await act(async () => {
-    await ref.current;
   });
 
   expect(insertedHTML).toHaveBeenLastCalledWith(
@@ -95,7 +91,7 @@ test('it applies provided nonce on the injected script tag', async () => {
 test('it waits for both search and recommend results', async () => {
   const ref: { current: PromiseWithState<void> | null } = { current: null };
 
-  const client = renderComponent({
+  const client = await renderComponent({
     ref,
     children: (
       <>
@@ -103,10 +99,6 @@ test('it waits for both search and recommend results', async () => {
         <TrendingItems />
       </>
     ),
-  });
-
-  await act(async () => {
-    await ref.current;
   });
 
   expect(ref.current!.status).toBe('fulfilled');
@@ -117,17 +109,13 @@ test('it waits for both search and recommend results', async () => {
 test('it waits for search only if there are only search widgets', async () => {
   const ref: { current: PromiseWithState<void> | null } = { current: null };
 
-  const client = renderComponent({
+  const client = await renderComponent({
     ref,
     children: (
       <>
         <SearchBox />
       </>
     ),
-  });
-
-  await act(async () => {
-    await ref.current;
   });
 
   expect(ref.current!.status).toBe('fulfilled');
@@ -141,17 +129,13 @@ test('it waits for search only if there are only search widgets', async () => {
 test('it waits for recommend only if there are only recommend widgets', async () => {
   const ref: { current: PromiseWithState<void> | null } = { current: null };
 
-  const client = renderComponent({
+  const client = await renderComponent({
     ref,
     children: (
       <>
         <TrendingItems />
       </>
     ),
-  });
-
-  await act(async () => {
-    await ref.current;
   });
 
   expect(ref.current!.status).toBe('fulfilled');